backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
redis==5.2.1
httpx==0.28.1
beautifulsoup4==4.12.3
google-generativeai==0.8.3
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.0.1
